Fix warnings from -Wsign-compare -Wreorder -Wcomment
-Wsign-compare. We were comparing size types with ints; fixed by recasting.

-Wreorder. The member order specified in the class definitions was not
           matching the initialization order.

-Wcomment. There was an extra /* sign within a block comment.

+++ block.h
@@ -101,7 +101,7 @@ public:
 
 
 	vector<int> get_needed_res(int length) {
-		if (length != res_needed.size()) {
+        if (length != (int)res_needed.size()) {
 			cout << "Integer provided to get_needed_res does not agree with resource vector length" << endl;
 		}
